Use std::size_t in Z_

+++ utils/memory.hpp
@@ -19,5 +19,5 @@ auto create_struct(const std::size_t size) {
 
 template <typename DataType>
 auto zmalloc(size_t size, int tag, void * ptr) {
-  return static_cast<DataType>(Z_Malloc(static_cast<int>(size), tag, ptr));
+  return static_cast<DataType>(Z_Malloc(size, tag, ptr));
 }
